% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-fullscreen.R
\name{addControlFullScreen}
\alias{addControlFullScreen}
\title{Add full screen control to map}
\usage{
addControlFullScreen(map)
}
\arguments{
\item{map}{the map to add the tile layer to}
}
\value{
modified map object
}
\description{
Add full screen control (https://github.com/brunob/leaflet.fullscreen)
}
\examples{
leaflet() \%>\%
  addControlFullScreen()
}

