% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-search.R
\name{addSearchOSM}
\alias{addSearchOSM}
\title{Add a OSM search control to the map.}
\usage{
addSearchOSM(map, position = "topright")
}
\arguments{
\item{map}{a map widget object}

\item{position}{standard \href{http://leafletjs.com/reference.html#control-positions}{Leaflet control position options}.}
}
\value{
modified map
}
\description{
Add a OSM search control to the map.
}
\examples{
leaflet(data = quakes) \%>\%
addTiles() \%>\%
    addSearchOSM()
}

